Badges actually work!
Global badges are now loaded once. The cache is not (yet) reset during
the lifetime of the application. This is to prevent querying Twitch for
data that hasn't been modified. This is accomplished via the new
initialize() function.

Added proper functions to the Cache library (add, set, get_maybe, etc).

Rearranged code to be more modular.

Fixed /badge/:set endpoint. It now works.

Added several test cases to ensure the application works as expected.